sections/10-client/requirements.txt
streamlit
snowflake-snowpark-python[pandas]
networkx
pyvis
plotly